(revert): removed role split, need to do more research (feat): introduced utils.lua file to shared for access from server & client (feat): Added transitions for buttons when leaving focus

+++ server/server.lua
@@ -1,17 +1,29 @@
+local QBCore = nil;
+local playerRoles = {}
+local utils = require('shared/utils')
+
+if Config.Framework == 'qbx' then
+    print('[LBS_ADMIN] Qbox detected')
+elseif Config.Framework == 'qb' then
+    QBCore = exports['qb-core']:GetCoreObject();
+    print('[LBS_ADMIN] QBCore detected')
+else   
+    print('^1[LBS_ADMIN] No supported framework detected. Ensure qbox or qbcore is running.^7')
+end
+
+
+
 
 ------------------------------
 -- FUNCTIONS
 ------------------------------
-
 function send_to_discord_log(title, description, color)
     local webhook = Config.DiscordWebhook
-    if (not webhook or webhook == '') and Config.EnableDebugMode then 
-        print("[Discord Log] No webhook URL set!")
+    if (not webhook or webhook == '') then 
+        utils.debugPrint("[Discord Log] No webhook URL set!")
         return 
     end
-    if Config.EnableDebugMode then 
-        print("[Discord Log] Sending to Discord:", title, description)
-    end 
+    utils.debugPrint("[Discord Log] Sending to Discord:", title, description)
 
     local embed = {
         {
@@ -34,11 +46,12 @@ function send_to_discord_log(title, description, color)
 
 end
 
-
 local function hasAdminPermissions(src)
-    local admin_role = Config.AdminGroup
-    local admin_status = IsPlayerAceAllowed(src, admin_role)
-    return admin_status or IsPlayerAceAllowed(src, 'command')
+    return IsPlayerAceAllowed(src, 'command')
+end
+
+local function hasPermission(src, perm)
+    return IsPlayerAceAllowed(src, perm) or IsPlayerAceAllowed(src, 'command')
 end
 
 local function getBanReason(reason, duration, durationUnit)
@@ -92,13 +105,12 @@ local function deleteBanRecord(identifier)
         return
     end
 
-    for key, value in pairs(jsonTable) do
-        print("Key: " .. key .. ", Value: " .. (type(value) == "table" and "Table" or tostring(value)))
-    end
 
     if jsonTable[identifier] then
         jsonTable[identifier] = nil
-        print("[DEBUG] Removed entry for identifier: " .. identifier)
+        if Config.EnableDebugMode then
+            print("[DEBUG] Removed entry for identifier: " .. identifier)
+        end
     else
         print("[ERROR] Identifier '" .. identifier .. "' not found in bans.json.")
         return
@@ -117,44 +129,36 @@ local function deleteBanRecord(identifier)
     SaveResourceFile(resourceName, filePath, reconstructedContent, -1)
     return true
 end
+
+
 ------------------------------
 --CALLBACKS
 ------------------------------
 lib.callback.register('lbs_admin:server:check_permissions', function(src)
-    local admin_role = Config.AdminGroup
-    local admin_status = IsPlayerAceAllowed(src, admin_role)
-    --should double check if they have command access ~_~
-    return admin_status or IsPlayerAceAllowed(src, 'command')
+    return IsPlayerAceAllowed(src, 'command')
 end)
 
 lib.callback.register('lbs_admin:server:getPlayerList', function()
     local players = {}
     if Config.Framework == 'qb' then 
-        for _, player in pairs(GetPlayers()) do
-            if Config.EnableDebugMode then
-                print(string.format("[DEBUG] Player found: %s [%s]", player.PlayerData.name, player.PlayerData.source))
-            end
+        for _, player in pairs(QBCore.Functions.GetQBPlayers()) do
+            utils.debugPrint(string.format("[DEBUG] Player found: %s [%s]", player.PlayerData.name, player.PlayerData.source))
             table.insert(players, {
                 id = player.PlayerData.source,
                 name = player.PlayerData.name or 'Unknown'
             })
         end
-        if Config.EnableDebugMode then
-            print('[DEBUG] Sending player list with ' .. #players .. ' players')
-        end
+        utils.debugPrint('[DEBUG] Sending player list with ' .. #players .. ' players')
     elseif Config.Framework == 'qbx' then
         for _, player in pairs(exports.qbx_core:GetQBPlayers()) do
-            if Config.EnableDebugMode then
-                print(string.format("[DEBUG] Player found: %s [%s]", player.PlayerData.name, player.PlayerData.source))
-            end
+            utils.debugPrint(string.format("[DEBUG] Player found: %s [%s]", player.PlayerData.name, player.PlayerData.source))
             table.insert(players, {
                 id = player.PlayerData.source, 
                 name = player.PlayerData.name or 'Unknown'
             })
         end
-        if Config.EnableDebugMode then
-            print('[DEBUG] Sending player list with ' .. #players .. ' players')
-        end
+        utils.debugPrint('[DEBUG] Sending player list with ' .. #players .. ' players')
+
     else 
         print('No Framework Found!')
     end
@@ -187,6 +191,8 @@ lib.callback.register('lbs_admin:server:getReportList', function()
                 status = row.status
             })
         end
+    else
+        utils.debugPrint('[DEBUG] Failed to query reports table in database')
     end
 
 
@@ -279,7 +285,6 @@ lib.callback.register('lbs_admin:server:getBansList', function()
     end
     return bans
 end)
-
 ------------------------------
 --Events
 ------------------------------
@@ -289,8 +294,7 @@ RegisterNetEvent('lbs_admin:server:report_action', function(action, target)
     local src = source
     if not target then return end
     if action == 'close' then
-        if hasAdminPermissions(src) then
-
+        if hasPermission(src, 'reports') then
             local queries = {
                 {query = 'DELETE FROM `reports` WHERE reporter_id = ?', values= {target}}
             }
@@ -324,7 +328,7 @@ RegisterNetEvent('lbs_admin:server:player_action', function(action, target, reas
     -- BAN ACTION
     if action == 'ban' then
         --log to discord
-        if hasAdminPermissions(source) then
+        if hasPermission(source, 'ban') then
             reason, banTime = getBanReason(reason, duration, durationUnit)
             send_to_discord_log("BAN Action", ("%s [%s] banned %s [%s] for: \n%s "):format(admin,source,player,target,reason), 255)
             MySQL.insert('INSERT INTO bans (name, license, discord, ip, reason, expire, bannedby) VALUES (?,?,?,?,?,?,?)',{
@@ -339,17 +343,21 @@ RegisterNetEvent('lbs_admin:server:player_action', function(action, target, reas
             DropPlayer(target, reason)
         end
     elseif action == 'teleport' then
-        local src = source 
-        local coords = GetEntityCoords(GetPlayerPed(target))
-        send_to_discord_log("Teleport Action", ("%s [%s] teleported to %s [%s]"):format(admin,source,player,target), 255)
-        TriggerClientEvent('lbs_admin:client:teleport_to_coords', src, coords)
+        if hasPermission(source, 'teleport') then 
+            local src = source 
+            local coords = GetEntityCoords(GetPlayerPed(target))
+            send_to_discord_log("Teleport Action", ("%s [%s] teleported to %s [%s]"):format(admin,source,player,target), 255)
+            TriggerClientEvent('lbs_admin:client:teleport_to_coords', src, coords)
+        end
     elseif action == 'bring' then
-        local src = source
-        local coords = GetEntityCoords(GetPlayerPed(src))
-        send_to_discord_log("Bring Action", ("%s [%s] teleported %s [%s]"):format(admin,source,player,target), 32896)
-        TriggerClientEvent('lbs_admin:client:teleport_to_coords', target, coords)
+        if hasPermission(source, 'teleport') then 
+            local src = source
+            local coords = GetEntityCoords(GetPlayerPed(src))
+            send_to_discord_log("Bring Action", ("%s [%s] teleported %s [%s]"):format(admin,source,player,target), 32896)
+            TriggerClientEvent('lbs_admin:client:teleport_to_coords', target, coords)
+        end
     elseif action == 'spectate' then
-        if hasAdminPermissions(source) then
+        if hasPermission(source, 'spectate') then
             if source == target then 
                 TriggerClientEvent('ox_lib:notify', source , {
                     title = 'Error',
@@ -364,7 +372,7 @@ RegisterNetEvent('lbs_admin:server:player_action', function(action, target, reas
             TriggerClientEvent('lbs_admin:client:spectate', source, targetPed)
         end
     elseif action == 'kick' then
-        if hasAdminPermissions(source) then
+        if hasPermission(source, 'kick') then
             send_to_discord_log("Kick Action", ("%s [%s] kicked %s [%s] for: %s"):format(admin,source,player,target, reason), 16711680)
             local discordLink = Config.DiscordLink
             DropPlayer(target, "You were kicked for: \n" .. reason .. "\nJoin our Discord for more information: " .. discordLink)
@@ -372,12 +380,13 @@ RegisterNetEvent('lbs_admin:server:player_action', function(action, target, reas
     end
 end)
 
-
 RegisterNetEvent('lbs_admin:server:teleport_marker', function()
     local src = source
-    if Config.Framework == 'qb' or Config.Framework == 'qbx' then
-        TriggerClientEvent('QBCore:Command:GoToMarker', src)
-        send_to_discord_log("TPM Action", ("%s [%s] teleport to marker"):format(admin,source), 255)
+    if hasPermission(src, 'teleport') then
+        if Config.Framework == 'qb' or Config.Framework == 'qbx' then
+            TriggerClientEvent('QBCore:Command:GoToMarker', src)
+            send_to_discord_log("TPM Action", ("%s [%s] teleport to marker"):format(admin,source), 255)
+        end
     end
 end)
 
@@ -397,8 +406,10 @@ RegisterNetEvent('lbs_admin:server:submitReport', function(message)
 end)
 
 RegisterNetEvent('lbs_admin:server:unbanPlayer', function(banId)
-    if not banId then return end
-    deleteBanRecord(banId)
-    local admin = GetPlayerName(source)
-    send_to_discord_log("Unban Action", ("%s [%s] unbanned ban ID: %s "):format(admin,source,banId), 255)
+    if hasPermission(source, 'ban') then 
+        if not banId then return end
+        deleteBanRecord(banId)
+        local admin = GetPlayerName(source)
+        send_to_discord_log("Unban Action", ("%s [%s] unbanned ban ID: %s "):format(admin,source,banId), 255)
+    end
 end)
